requirements.txt
appdirs==1.4.4
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
colorama==0.4.1
docutils==0.15.2
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==0.22
isort==4.3.21
jdcal==1.4.1
Kivy==1.11.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
openpyxl==3.0.0
packaging==19.1
pandas==0.25.1
pathspec==0.8.0
pluggy==0.13.0
py==1.8.0
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.3.1
regex==2020.5.14
requests==2.22.0
six==1.12.0
toml==0.10.1
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
xlrd==1.2.0
zipp==0.6.0
